requirements.txt
click==8.0.1
cycler==0.10.0
Flask==2.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mysql-connector-python==8.0.25
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
